refactor: moves helper functions to utils
